MS-42 Other minor cleanup in events DB config
