+++ infra/events/src/main/java/com/simprints/infra/events/event/local/SessionScopeRoomDao.kt
@@ -10,20 +10,11 @@ import com.simprints.infra.events.event.local.models.DbEventScope
 @Dao
 internal interface SessionScopeRoomDao {
 
-    @Query("select * from DbEventScope where end_unixMs IS NULL order by start_unixMs desc")
-    suspend fun loadOpen(): List<DbEventScope>
-
     @Query("select * from DbEventScope where type = :type AND end_unixMs IS NULL order by start_unixMs desc")
     suspend fun loadOpen(type: EventScopeType): List<DbEventScope>
 
-    @Query("select * from DbEventScope where end_unixMs IS NOT NULL order by start_unixMs desc")
-    suspend fun loadClosed(): List<DbEventScope>
-
     @Query("select * from DbEventScope where type = :type AND end_unixMs IS NOT NULL order by start_unixMs desc")
     suspend fun loadClosed(type: EventScopeType): List<DbEventScope>
-
-    @Query("select count(*) from DbEventScope")
-    suspend fun count(): Int
 
     @Query("select count(*) from DbEventScope where type = :type")
     suspend fun count(type: EventScopeType): Int
