+++ infra/events/src/test/java/com/simprints/infra/events/event/local/EventLocalDataSourceTest.kt
@@ -10,6 +10,7 @@ import com.google.common.truth.Truth.assertThat
 import com.simprints.core.tools.json.JsonHelper
 import com.simprints.infra.events.event.domain.models.Event
 import com.simprints.infra.events.event.domain.models.EventType.CALLBACK_ENROLMENT
+import com.simprints.infra.events.event.domain.models.scope.EventScopeType
 import com.simprints.infra.events.event.local.*
 import com.simprints.infra.events.event.local.models.DbEvent
 import com.simprints.infra.events.event.local.models.DbEventScope
@@ -36,7 +37,7 @@ import java.io.IOException
 
 @RunWith(AndroidJUnit4::class)
 @Config(application = HiltTestApplication::class, shadows = [ShadowAndroidXMultiDex::class])
-internal class EventLocalDataSourceImplTest {
+internal class EventLocalDataSourceTest {
 
     private lateinit var db: EventRoomDatabase
     private lateinit var eventDao: EventRoomDao
@@ -60,7 +61,7 @@ internal class EventLocalDataSourceImplTest {
         every { eventDatabaseFactory.build() } returns db
         mockDaoLoadToMakeNothing()
 
-        eventLocalDataSource = EventLocalDataSourceImpl(
+        eventLocalDataSource = EventLocalDataSource(
             eventDatabaseFactory,
             JsonHelper,
             UnconfinedTestDispatcher(),
@@ -346,9 +347,9 @@ internal class EventLocalDataSourceImplTest {
 
     @Test
     fun countSessions() = runTest {
-        eventLocalDataSource.countEventScopes()
+        eventLocalDataSource.countEventScopes(EventScopeType.SESSION)
 
-        coVerify { scopeDao.count() }
+        coVerify { scopeDao.count(EventScopeType.SESSION) }
     }
 
     @Test
@@ -365,10 +366,10 @@ internal class EventLocalDataSourceImplTest {
         val dbSessionCaptureEvent = mockk<DbEventScope> {
             every { fromDbToDomain(any()) } returns mockk()
         }
-        coEvery { scopeDao.loadOpen() } returns listOf(dbSessionCaptureEvent)
-        eventLocalDataSource.loadOpenedScopes()
+        coEvery { scopeDao.loadOpen(EventScopeType.SESSION) } returns listOf(dbSessionCaptureEvent)
+        eventLocalDataSource.loadOpenedScopes(EventScopeType.SESSION)
 
-        coVerify { scopeDao.loadOpen() }
+        coVerify { scopeDao.loadOpen(any()) }
         verify { dbSessionCaptureEvent.fromDbToDomain(any()) }
     }
 
@@ -378,10 +379,10 @@ internal class EventLocalDataSourceImplTest {
         val dbSessionCaptureEvent = mockk<DbEventScope> {
             every { fromDbToDomain(any()) } returns mockk()
         }
-        coEvery { scopeDao.loadClosed() } returns listOf(dbSessionCaptureEvent)
-        eventLocalDataSource.loadClosedScopes()
+        coEvery { scopeDao.loadClosed(EventScopeType.SESSION) } returns listOf(dbSessionCaptureEvent)
+        eventLocalDataSource.loadClosedScopes(EventScopeType.SESSION)
 
-        coVerify { scopeDao.loadClosed() }
+        coVerify { scopeDao.loadClosed(any()) }
         verify { dbSessionCaptureEvent.fromDbToDomain(any()) }
     }
 
@@ -457,8 +458,9 @@ internal class EventLocalDataSourceImplTest {
         eventDatabaseFactory = mockk(relaxed = true)
         coEvery { eventDao.loadAll() } returns emptyList()
         coEvery { eventDao.loadFromSession(any()) } returns emptyList()
-        coEvery { scopeDao.loadOpen() } returns emptyList()
-        coEvery { scopeDao.count() } returns 0
+        coEvery { scopeDao.loadOpen(any()) } returns emptyList()
+        coEvery { scopeDao.loadClosed(any()) } returns emptyList()
+        coEvery { scopeDao.count(any()) } returns 0
         every { db.eventDao } returns eventDao
         every { db.scopeDao } returns scopeDao
         every { eventDatabaseFactory.build() } returns db
